Move the function isSensorReady to LibreUtils.
